refactor(invitations): reorganize files
